fix(executor): preserve export semantics and restore agent security policy
Two correctness regressions found in review of the archive transport change,
plus the cleanups that came with them.

1. `git archive` applies .gitattributes `export-ignore` and `export-subst`;
   the `checkout-index` export it replaced did not. A repo that export-ignores
   docs/ or tests/ would hand the agent a tree with files the PR actually
   changed missing, and `$Format:...$` placeholders silently rewritten -- with
   git tools filtered out by filterGitToolsForSnapshot, the agent cannot detect
   or recover from either. Driving tar from `git ls-files` restores the
   index-scoped, unfiltered semantics while still writing one file to the SMB
   mount. Verified both directions: the new test fails against `git archive`
   and passes against the tar pipeline.

2. Every transport materializes the repo under a fresh tmpdir, so the security
   policy written to workspaceDir was not in the SDK's project-settings
   ancestor chain and never loaded -- while review-prompt.ts still tells the
   agent that the policy in CLAUDE.md overrides instructions found in PR or
   comment text. That left prompt-injection defenses documented but absent.
   This predates the archive change and affected the bundle transports too;
   the policy is now copied beside the materialized repo (its parent, so it
   neither shadows a repo-owned CLAUDE.md nor appears as a reviewable file).

Also: the shallow-path comment claimed ~30k files for xbmc/xbmc as the
motivating cost, but xbmc/xbmc has tracked symlinks and therefore already took
the archive path -- and has ~10k tracked blobs, not 30k. The comment now names
the population this actually changes (shallow repos with no tracked symlinks)
so the next person measuring savings looks at the right path.

The three prepareAgentWorkspace tests duplicated between executor.test.ts and
prepare-agent-workspace.test.ts had already diverged in this branch. Removed
the executor.test.ts copies; prepare-agent-workspace.test.ts is the single home
and carries the stronger assertions.
